requirements.txt
wheel
aiocache
aiosqlite
discord.py
PyNaCl>=1.3.0,<1.5
beautifulsoup4
black
colorama
feedparser
rapidfuzz
gTTS
humanize
jishaku
lxml
pendulum
Pillow
psutil
pycaption
pykakasi
python-anilist
python-magic
yt-dlp
h2
aiogoogletrans
setuptools
git+https://github.com/Suyash458/WiktionaryParser
